Cleaned up code and seperated code from spotify routes that had nothing to do with routes
